Reduce amount of ladders and logs
Upon complaints noticed, im proposing the amount of ladders received upon spawning is 4 and that log is 16. This is because inventories are filling up very fast for many players, and most of these payers dont even use the ladders and logs anyway. also made it so breaking diamond blocks turns it into gold blocks.