Change of large amount of game mechanics
Changed iron ingot to gold ingot, so it is possible to make gold armor. Iron armor was argued to be to over powered. Because of this change, I have also decided to change the stone sword to an iron sword. Changed item lore and name of stone sword to apply it to iron sword.